chore: move 'script_kind' under its own module.
